Module Grpc_discovery

A simple service discovery via file. This supports both unix socket and tcp servers running on localhost.

module Discovery_file : sig ... end

Encoding a sockaddr into a file.

module Connection_config : sig ... end

The client side of the discovery answers the question: "Where is the service running?"

module Listening_config : sig ... end

The server side of the discovery specifies where to serve, and how to advertize that information so clients can find you.