Module Grpc_discovery.Connection_config

The client side of the discovery answers the question: "Where is the service running?"

The intended usage for this library is to add param to you command line parameters, and resolve the t using sockaddr in the body of your client command.

type t =

| Tcp of {

host : [ `Localhost | `Ipaddr of Eio.Net.Ipaddr.v4v6 ];

port : Base.int;

}

| Unix of {

path : Fpath.t;

}

| Discovery_file of {

path : Fpath.t;

}

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

val param : t Command.Param.t

val sockaddr : t -> env:< fs : [> Eio.Fs.dir_ty ] Eio.Path.t.. > -> Eio.Net.Sockaddr.stream Base.Or_error.t

val to_params : t -> Base.string Base.list

Returns the arguments that a client command needs to be supplied to rebuild t via param. This is used by tests and by Grpc_test.Config.grpc_discovery to create the right invocations for clients whose cli uses param.